Fix/prep staging migration (#118)
* add distinct

* add module_name

* fix MRO stuff in migration

* fix project cache rebuild

* fix new cmd and use it in migration

* Update project/models/create.py


---------
