+++ project/management/commands/build_project_ocsge_status.py
@@ -28,6 +28,8 @@ class Command(BaseCommand):
             logger.info("Filtering on departements: %s", options["departements"])
             projects = projects.filter(cities__departement__source_id__in=options["departements"])
 
+        projects = projects.distinct()
+
         count = projects.count()
 
         logger.info(f"{count} projects")
