+++ public_data/models/ocsge.py
@@ -18,28 +18,15 @@ Remarque : il est possible qu'il faille partitionner ces données à terme pour 
 problème de performance
 
 """
-from typing import Self
-
 from django.contrib.gis.db import models
 from django.contrib.gis.db.models.functions import Area, Intersection, MakeValid
 from django.core.validators import MaxValueValidator, MinValueValidator
-from django.db.models import DecimalField, Sum
-from django.db.models.functions import Cast
+from django.db.models import Sum
 
 from public_data.models.couverture_usage import CouvertureUsageMatrix
 from public_data.models.enums import SRID
-from public_data.models.mixins import (
-    AutoLoadMixin,
-    DataColorationMixin,
-    TruncateTableMixin,
-)
+from public_data.models.mixins import DataColorationMixin, TruncateTableMixin
 from utils.db import DynamicSRIDTransform, IntersectManager
-
-
-# syntaxic sugar to avoid writing long line of code
-# cache has been added to matrix_dict method
-def get_matrix(cs, us):
-    return CouvertureUsageMatrix().matrix_dict()[(cs, us)]
 
 
 class Ocsge(TruncateTableMixin, DataColorationMixin, models.Model):
@@ -96,52 +83,6 @@ class Ocsge(TruncateTableMixin, DataColorationMixin, models.Model):
         return data
 
 
-class AutoOcsge(AutoLoadMixin, Ocsge):
-    class Meta:
-        proxy = True
-
-    mapping = {
-        "id_source": "ID",
-        "couverture": "CODE_CS",
-        "usage": "CODE_US",
-        "mpoly": "MULTIPOLYGON",
-    }
-
-    def save(self, *args, **kwargs) -> Self:
-        self.year = self.__class__._year
-        self.departement = self.__class__._departement
-        self.srid_source = self.srid
-
-        self.matrix = get_matrix(self.couverture, self.usage)
-        self.is_artificial = bool(self.matrix.is_artificial)
-
-        if self.matrix.couverture:
-            self.couverture_label = self.matrix.couverture.label
-        if self.matrix.usage:
-            self.usage_label = self.matrix.usage.label
-
-        return super().save(*args, **kwargs)
-
-    @classmethod
-    def clean_data(cls) -> None:
-        cls.objects.filter(
-            departement=cls._departement,
-            year=cls._year,
-        ).delete()
-
-    @classmethod
-    def calculate_fields(cls) -> None:
-        cls.objects.filter(
-            departement=cls._departement,
-            year=cls._year,
-        ).update(
-            surface=Cast(
-                Area(DynamicSRIDTransform("mpoly", "srid_source")),
-                DecimalField(max_digits=15, decimal_places=4),
-            )
-        )
-
-
 class OcsgeDiff(TruncateTableMixin, DataColorationMixin, models.Model):
     year_old = models.IntegerField("Ancienne année", validators=[MinValueValidator(2000), MaxValueValidator(2050)])
     year_new = models.IntegerField("Nouvelle année", validators=[MinValueValidator(2000), MaxValueValidator(2050)])
@@ -190,69 +131,6 @@ class OcsgeDiff(TruncateTableMixin, DataColorationMixin, models.Model):
             models.Index(fields=["year_old"]),
             models.Index(fields=["year_new"]),
         ]
-
-
-class AutoOcsgeDiff(AutoLoadMixin, OcsgeDiff):
-    class Meta:
-        proxy = True
-
-    @classmethod
-    @property
-    def mapping(cls) -> dict[str, str]:
-        return {
-            "cs_old": f"CS_{cls._year_old}",
-            "us_old": f"US_{cls._year_old}",
-            "cs_new": f"CS_{cls._year_new}",
-            "us_new": f"US_{cls._year_new}",
-            "mpoly": "MULTIPOLYGON",
-        }
-
-    def before_save(self) -> None:
-        self.year_new = self.__class__._year_new
-        self.year_old = self.__class__._year_old
-        self.departement = self.__class__._departement
-        self.srid_source = self.srid
-
-        self.new_matrix = get_matrix(self.cs_new, self.us_new)
-        self.new_is_artif = bool(self.new_matrix.is_artificial)
-
-        if self.new_matrix.couverture:
-            self.cs_new_label = self.new_matrix.couverture.label
-
-        if self.new_matrix.usage:
-            self.us_new_label = self.new_matrix.usage.label
-
-        self.old_matrix = get_matrix(self.cs_old, self.us_old)
-        self.old_is_artif = bool(self.old_matrix.is_artificial)
-
-        if self.old_matrix.couverture:
-            self.cs_old_label = self.old_matrix.couverture.label
-        if self.old_matrix.usage:
-            self.us_old_label = self.old_matrix.usage.label
-
-        self.is_new_artif = not self.old_is_artif and self.new_is_artif
-        self.is_new_natural = self.old_is_artif and not self.new_is_artif
-
-    @classmethod
-    def calculate_fields(cls) -> None:
-        cls.objects.filter(
-            departement=cls._departement,
-            year_new=cls._year_new,
-            year_old=cls._year_old,
-        ).update(
-            surface=Cast(
-                Area(DynamicSRIDTransform("mpoly", "srid_source")),
-                DecimalField(max_digits=15, decimal_places=4),
-            )
-        )
-
-    @classmethod
-    def clean_data(cls) -> None:
-        cls.objects.filter(
-            departement=cls._departement,
-            year_new=cls._year_new,
-            year_old=cls._year_old,
-        ).delete()
 
 
 class ArtificialArea(TruncateTableMixin, DataColorationMixin, models.Model):
@@ -309,29 +187,3 @@ class ZoneConstruite(TruncateTableMixin, DataColorationMixin, models.Model):
         indexes = [
             models.Index(fields=["year"]),
         ]
-
-
-class AutoZoneConstruite(AutoLoadMixin, ZoneConstruite):
-    class Meta:
-        proxy = True
-
-    mapping = {
-        "id_source": "ID",
-        "millesime": "MILLESIME",
-        "mpoly": "MULTIPOLYGON",
-    }
-
-    def save(self, *args, **kwargs) -> Self:
-        self.year = int(self._year)
-        self.departement = self.__class__._departement
-        self.srid_source = self.srid
-        self.surface = self.mpoly.transform(self.srid, clone=True).area
-        self.departement = self._departement
-        return super().save(*args, **kwargs)
-
-    @classmethod
-    def clean_data(cls) -> None:
-        cls.objects.filter(
-            departement=cls._departement,
-            year=cls._year,
-        ).delete()
